Correct

+++ LoanerDemo/Classes/OCFile/ZLPhotoLib/ZLPhotoBrowser/ViewControllers/ZLPhotoPickerBrowserViewController.m
@@ -454,7 +454,7 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 }
 
 -(void)setPageLabelPage:(NSInteger)page{
-    self.pageLabel.text = [NSString stringWithFormat:@"%ld / %ld",page + 1, self.photos.count];
+    self.pageLabel.text = [NSString stringWithFormat:@"%ld / %ld",page + 1, (unsigned long)self.photos.count];
 }
 
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView{
